# api/openapi.yaml
components:
  schemas:
...
    AmrEntry:
      description: One native authenticator proof - which authenticator TYPE, and which specific execution/instance of it.
      properties:
        amrSourceId:
          example: kc-otp-form-exec-1
          type: string
        nativeToolId:
          example: kc-otp-form
          type: string
      required:
      - amrSourceId
      - nativeToolId
      type: object
...
    AuthData:
      properties:
        acr:
          example: loa2
          type: string
        amr:
          additionalProperties:
            type: string
          description: 'Method -> who proved it: "orchestrator" for a completed orchestrator tool, "kc" for evidence a native
            Keycloak authenticator already established (docs/05-api.md Abschnitt 3). Informational only - the orchestrator
            alone still resolves the combined acr above, regardless of source.'
          example:
            password: kc
            sms: orchestrator
          type: object
        subject:
          $ref: '#/components/schemas/AuthSubject'
      type: object
...
    AuthSubject:
      properties:
        id:
          example: '42'
          type: string
        type:
          enum:
          - account
          - invitation
          type: string
      required:
      - id
      - type
      type: object
...
    ChannelBlock:
      properties:
        activeMethods:
          description: All active authentication methods on the account, regardless of whether this session's currentAmr proved
            them. currentAmr is session evidence (what THIS channel actually proved); activeMethods is the account's full
            standing method list, unfiltered by device - a lost/stolen device's credential must be removable from any authenticated
            session, not only from that device itself.
          items:
            $ref: '#/components/schemas/ActiveMethodView'
          type: array
        channelSessionId:
          format: uuid
          type: string
        channelType:
          description: Which facade this channel was opened through - APP (DPoP) or WEB (docs/02-domaenenmodell.md Abschnitt
            1). Fixed for the channel's whole lifetime.
          example: APP
          type: string
        currentAcr:
          example: loa2
          type: string
        currentAmr:
          example:
          - sms
          - password
          items:
            type: string
          type: array
        hasProvenFactor:
          description: Whether this session has already proven at least one factor (an identification or a method) - what
            cancelling the running journey would throw away. Lets a client ask "really discard?" only when there is something
            to lose. Set on every response, tool responses included; says nothing about which factor.
          example: false
          type: boolean
        state:
          example: AUTHENTICATED
          type: string
      required:
      - channelSessionId
      - channelType
      - state
      type: object
...
    ChannelResponse:
      properties:
        authData:
          $ref: '#/components/schemas/AuthData'
          description: WEB channels only (docs/05-api.md Abschnitt 3) - never present for APP.
        channel:
          $ref: '#/components/schemas/ChannelBlock'
        demo:
          $ref: '#/components/schemas/DemoInfo'
          description: Demo-only correlation IDs, never part of the production contract.
        next:
          $ref: '#/components/schemas/Next'
        stepData:
          $ref: '#/components/schemas/StepData'
          description: Whatever the current step needs to render. `kind` names the shape - see StepData.
      required:
      - channel
      type: object
...
    DemoInfo:
      additionalProperties: true
      properties:
        accountId:
          example: 42
          format: int64
          type: integer
        journeys:
          description: The running journey chain for this channel, outermost first - see [JourneyDebugStep]. Empty once nothing
            is running.
          items:
            $ref: '#/components/schemas/JourneyDebugStep'
          type: array
        personId:
          example: P000000001
          type: string
        session:
          $ref: '#/components/schemas/DemoSession'
          description: 'Demo-only: who this session belongs to and what it has proven so far - in every response, tool responses
            included, so a demo view can show it at any step. The production contract keeps these on the channel resource
            only (ChannelBlock).'
      type: object
    DemoSession:
      properties:
        acr:
          example: loa1
          type: string
        amr:
          example:
          - sms
          items:
            type: string
          type: array
        authenticated:
          description: Whether the session is signed in (channel state AUTHENTICATED).
          type: boolean
        personName:
          description: '"Vorname Name" of the person behind the account, if one is bound.'
          example: Max Muster
          type: string
      required:
      - authenticated
      type: object
...
    ErrorResponse:
      description: |-
        Every error response has this shape. The HTTP status is fixed per `error` code:

        - `BAD_REQUEST`: 400
        - `UNAUTHORIZED`: 401
        - `BINDING_MISMATCH`: 403
        - `NOT_FOUND`: 404
        - `INVALID_STATE_TRANSITION`: 409
        - `CONCURRENT_MODIFICATION`: 409
        - `PROCESS_GONE`: 410
        - `PROCESS_ABORTED`: 410
        - `UNRESOLVABLE_REFERENCE`: 422
        - `ACCOUNT_LOCKED`: 423
        - `TOO_MANY_REQUESTS`: 429
        - `INTERNAL_ERROR`: 500

        A client must expect a code it does not know and handle it by its HTTP status.
      properties:
        error:
          enum:
          - BAD_REQUEST
          - UNAUTHORIZED
          - BINDING_MISMATCH
          - NOT_FOUND
          - INVALID_STATE_TRANSITION
          - CONCURRENT_MODIFICATION
          - PROCESS_GONE
          - PROCESS_ABORTED
          - UNRESOLVABLE_REFERENCE
          - ACCOUNT_LOCKED
          - TOO_MANY_REQUESTS
          - INTERNAL_ERROR
          type: string
        text:
          $ref: '#/components/schemas/TextRef'
      required:
      - error
      - text
...
    KcChannelUpsertRequest:
      description: Upsert body for the kc-facade's one facade-specific endpoint (docs/05-api.md Abschnitt 3). All fields are
        optional. subject is whom Keycloak already knows (sub vorhanden) - an account binds the channel immediately, once,
        never overwritten by a later call. targetAcr is Keycloak's requested LoA level, already translated into an orchestrator
        ACR string, and only raises the channel's floor, never lowers it. amr lists which native Keycloak authenticators (never
        orchestrator tools) just proved something THIS flow run, one entry per proof - method/loa/factorTypes are resolved
        server-side from a NativeAuthenticatorDescriptor (see AmrEntry), the kc-facade's own mirror of a ToolDescriptor, not
        resolved from the orchestrator's own catalog (which stays entirely ignorant of native authenticators). Merged into
        the channel's evidence and re-checked against the current floor exactly like any other proof; no separate 'combined
        native acr' field exists, since the orchestrator derives that itself.
      properties:
        amr:
          items:
            $ref: '#/components/schemas/AmrEntry'
          type: array
        availableTools:
          description: The Web channel's own declaration of which toolIds its Keycloak theme can render (one com.example.identity.kcext.webtool.WebToolRenderer
            factory per toolId, registered via META-INF/services) - the kc-facade's counterpart to the App channel's own availableTools
            (POST /channels). Only read on this channel's first call (a later upsert resumes the already-persisted set); a
            channel-anonymous caller that omits this gets none of the orchestrator's tools, never all of them.
          items:
            type: string
          type: array
        intent:
          description: 'Only read on this channel''s first call, same restriction as availableTools - the kc facade''s own,
            deliberately narrow counterpart to the App facade''s `intent` request parameter (docs/05-api.md #"POST /app/channels:
            intent-Parameter"). Omitted (or null) means web_select_method, the existing login/step-up behaviour. Only web_select_method
            and register are accepted here - unlike the App facade, not every AuthIntent.isEntryIntent value: fast_access/lookup_login
            assume an APP-shaped channel this facade never has.'
          example: register
          type: string
        kcSessionId:
          description: Required whenever restoreData is present, ignored otherwise. Keycloak's own, durable UserSessionModel
            id - deliberately NOT read off the peer-auth assertion (the assertion's kc binding is always THIS flow run's own
            channelSessionId, docs/02-domaenenmodell.md Abschnitt 1, so it can't verify a token minted for a DIFFERENT, earlier
            flow run's channel). Must match what GET .../restore-data was called with to produce this exact restoreData token.
          type: string
        restoreData:
          description: A signed RestoreData token this same UserSession's channel returned earlier via GET .../restore-data,
            resubmitted verbatim (docs/05-api.md, section 3) - the bulk, one-shot way to seed a brand-new channel with what
            a PRIOR, unrelated flow run already established, as opposed to subject/amr above which report what THIS flow run
            just proved. Both are merged into the channel the same way; only restoreData may already be meaningfully old by
            the time it arrives here. Opaque to every caller but the orchestrator itself - see RestoreDataCodec.
          type: string
        subject:
          $ref: '#/components/schemas/AuthSubject'
          description: 'Whom Keycloak knows this flow run belongs to: an account or an invitation (ADR-48), the same shape
            as authData.subject in the answer. A channel already bound to another subject refuses it (409); an account binds
            a channel that has none, an invitation never does - only its own proof binds it.'
        targetAcr:
          example: loa2
          type: string
      type: object
    KcInvitationView:
      properties:
        attributes:
          additionalProperties:
            type: string
          type: object
        enabled:
          type: boolean
        firstName:
          type: string
        invitation:
          type: string
        lastName:
          type: string
        username:
          type: string
      required:
      - attributes
      - enabled
      - firstName
      - invitation
      - lastName
      - username
      type: object
...
    Next:
      properties:
        context:
          example: authentication
          type: string
        step:
          example: auth
          type: string
        toolId:
          example: auth-sms
          type: string
        toolSessionId:
          format: uuid
          type: string
        type:
          example: tool
          type: string
      required:
      - step
      - type
      type: object
...
    RestoreDataResponse:
      description: The channel's current RestoreData, signed - null if there is nothing worth restoring yet.
      properties:
        restoreData:
          type: string
      type: object
...
    TextRef:
      description: 'A text reference: look `key` up in the texts bundle (GET .../texts/{lang}), fill `{name}` placeholders
        from `args` (as is) or `texts` (resolved the same way, several joined with ", ").'
      properties:
        args:
          additionalProperties:
            type: string
          type: object
        key:
          example: account-not-found-08a2ef
          type: string
        template:
          description: The developer's wording, sent only while the texts bundle has none for key yet - show it instead of
            the key.
          type: string
        texts:
          additionalProperties:
            items:
              $ref: '#/components/schemas/TextRef'
            type: array
          type: object
      required:
      - key
...
    ToolCatalogEntry:
      properties:
        method:
          example: sms
          type: string
        role:
          example: KNOWN_ACCOUNT_AUTH
          type: string
        toolId:
          example: auth-sms
          type: string
      required:
      - method
      - role
      - toolId
      type: object
  securitySchemes:
    dpop:
      description: DPoP proof JWT, one per request, bound to method and URL (docs/09-dpop.md). The App channel's proof; the
        binding key is derived from it and is never supplied by the caller.
      in: header
      name: DPoP
      type: apiKey
    kc-peer-auth:
      bearerFormat: JWT
      description: 'Signed peer-auth assertion (docs/05-api.md Abschnitt 3, ADR-7). The Web channel''s proof: Keycloak calls
        server-to-server, and the channel binding comes out of the assertion.'
      scheme: bearer
      type: http
info:
  description: App-facing (orchestrator-first) API for the DPoP-bound registration and login demo. See docs/05-api.md for
    the full contract; the Keycloak-facing kc facade is out of scope for this build (docs/08-projektrahmen.md).
  title: Identity-Demo Orchestrator API
  version: v1
openapi: 3.1.0
paths:
...
  /orchestrator/api/v1/channels/{channelSessionId}:
    delete:
      description: 'Ends this channel for good (docs/02-domaenenmodell.md #3: AUTHENTICATED -> LOGGED_OUT, terminal) - cancels
        any active process and discards the AppTokenSession. For interactive clients prefer POST .../logouts which shows a
        confirmation prompt first. Never resumes on this channelSessionId afterwards; call POST .../channels again for a new
        session (a known device still skips straight to LOGIN there).'
      operationId: logout
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Logged out - no body.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Log out (direct, no confirmation)
      tags:
      - Channels
    get:
      description: 'The guaranteed resume entry point (docs/05-api.md #2): next always reflects the currently due step.'
      operationId: getChannel
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-sms
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Resumed mid-step-up, waiting on an SMS TAN.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - dpop: [
          ]
      - kc-peer-auth: [
          ]
      summary: Read the current channel state
      tags:
      - Channels
...
  /orchestrator/api/v1/kc/channels/{channelSessionId}:
    patch:
      description: 'Upsert semantics (docs/05-api.md Abschnitt 3): creates the channel on first call under this Keycloak-chosen
        id, just resumes it on every later one - idempotent by construction, no separate create-then-resume round trip. accountId/targetAcr
        are only meaningful on step-up.'
      operationId: upsertChannel
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KcChannelUpsertRequest'
      responses:
        '200':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  context: auth
                  step: selectMethod
                  type: orchestrator
                stepData:
                  kind: select-method
                  options:
                  - ident-fsc
                  - auth-password
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: First call for a fresh Keycloak flow run - offers every kc-usable method.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - kc-peer-auth: [
          ]
      summary: Upsert a kc channel and advance its journey
      tags:
      - KC channels
  /orchestrator/api/v1/kc/channels/{channelSessionId}/restore-data:
    get:
      description: For the Authenticator's end-of-flow lifecycle hook only (docs/ideen/docs/05-api.md Abschnitt 3) - reads
        back what this channel accumulated, to stash in a Keycloak UserSessionModel note and resubmit at a later flow's start.
        kcSessionId is passed explicitly (not read off the channel's own binding) because the returned token must remain valid
        across the flow-run boundary the channel itself does not survive - a fresh channel next flow run, but the very same
        UserSessionModel id. sessionExpiresAt (epoch seconds) is the latest end of that Keycloak session without further activity;
        the channel's expiry is capped at it (docs/adr/ADR-043).
      operationId: getRestoreData
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: kcSessionId
        required: true
        schema:
          type: string
      - in: query
        name: sessionExpiresAt
        required: false
        schema:
          format: int64
          type: integer
      - in: header
        name: Authorization
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestoreDataResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - kc-peer-auth: [
          ]
      summary: Fetch this channel's signed RestoreData
      tags:
      - KC channels
  /orchestrator/api/v1/kc/invitations/{invitation}:
    get:
      operationId: byInvitation
      parameters:
      - in: path
        name: invitation
        required: true
        schema:
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/KcInvitationView'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      security:
      - kc-peer-auth: [
          ]
      summary: The invitation as Keycloak shows it, as a user of its own
      tags:
      - KC invitation lookup
...
  /orchestrator/api/v1/tools/catalog:
    get:
      description: No auth, no channel required - purely descriptive.
      operationId: catalog
      responses:
        '200':
          content:
            application/json:
              example:
              - method: fsc
                role: IDENTIFICATION
                toolId: ident-fsc
              - method: sms
                role: ENROLLMENT
                toolId: enroll-sms
              - method: sms
                role: KNOWN_ACCOUNT_AUTH
                toolId: auth-sms
              - method: sms
                role: ACCOUNT_LOOKUP_AUTH
                toolId: auth-sms-lookup
              schema:
                items:
                  $ref: '#/components/schemas/ToolCatalogEntry'
                type: array
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error. `error` names the case; its HTTP status is fixed per code (see ErrorResponse).
      summary: List every registered tool
      tags:
      - Tool catalog
...